% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anovaBFcalc.R
\name{anovaBFcalc}
\alias{anovaBFcalc}
\title{A simple shiny application for computing Bayes factors from ANOVA summary statistics}
\usage{
anovaBFcalc(run = T, report_Apppath = F)
}
\arguments{
\item{run}{run the anovaBFcalc app (or not). Default is TRUE, meaning that the app will be lanuched.}

\item{report_Apppath}{report the directory of the app in your system. Default is FALSE, which will show the online app link.}
}
\description{
A simple shiny application for computing Bayes factors from ANOVA summary statistics
}
\note{
anovaBFcalc can directly be used on-line. Check See Also section.
}
\examples{
if (interactive()) {
  anovaBFcalc::anovaBFcalc()
}
}
\seealso{
Online version: \url{https://tomfaulkenberry.shinyapps.io/anovaBFcalc}

Tutorial : \url{https://github.com/tomfaulkenberry/anovaBFcalc}
}
\author{
Thomas J. Faulkenberry <faulkenberry@tarleton.edu>
}
