% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_bic.R
\name{bf_bic}
\alias{bf_bic}
\title{Compute BIC Bayes factor from ANOVA summary statistics.}
\usage{
bf_bic(F, df1, df2, repeated = FALSE, report.as = "BF01")
}
\arguments{
\item{F}{the observed F statistic}

\item{df1}{the between-treatments degrees of freedom}

\item{df2}{the residual degrees of freedom}

\item{repeated}{a logical indicating whether design is repeated
measures (TRUE) or between subjects (FALSE). Defaults to FALSE}

\item{report.as}{a string indicating whether to report Bayes factor
as support for null ("BF01") or alternative ("BF10"). Defaults
to "BF01"}
}
\description{
This function employs the BIC method (Kass and Raftery, 1995) 
to compute an approximation of the Bayes factor.
}
\details{
Minimally, the user must provide three inputs:
F = the observed F statistic
df1 = the "between-treatments" degrees of freedom
df2 = the residual degrees of freedom

The function outputs the value of BF_01 by default,
though this can be changed to BF_10 by the user (see
below).
}
\examples{
## using BIC approximation
## between subjects design, observed F(3,96) = 2.76
bf_bic(F=2.76, df1=3, df2=96)

## repeated measures design, observed F(1,22) = 1.336
bf_bic(F=1.336, df1=1, df2=22, repeated=TRUE)

## between subjects design, observed F(2,25) = 5.43
## report as support for alternative
bf_bic(F=5.43, df1=2, df2=25, report.as="BF10")
}
\references{
Kass, R. E., & Raftery, A. E. (1995). Bayes Factors. 
    Journal of the American Statistical Association, 90(430), 773. doi:10.2307/2291091
}
\author{
Thomas J. Faulkenberry <faulkenberry@tarleton.edu>
}
