% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anovaBFcalc.R
\name{anovaBFcalc}
\alias{anovaBFcalc}
\title{An interactive Shiny application for computing Bayes factors from ANOVA summary statistics}
\usage{
anovaBFcalc(run = T, report_path = F)
}
\arguments{
\item{run}{run the interactive app (or not). Default is TRUE, meaning that the app will be lanuched.}

\item{report_path}{report the directory of the app in your system. Default is FALSE, which will show the online app link.}
}
\description{
An interactive Shiny application for computing Bayes factors from ANOVA summary statistics
}
\note{
The interactive app can be directly accessed on-line. Check See Also section.
}
\examples{
if (interactive()) {
  anovaBFcalc::anovaBFcalc()
}
}
\seealso{
Online version: \url{https://tomfaulkenberry.shinyapps.io/anovaBFcalc}
}
\author{
Thomas J. Faulkenberry <faulkenberry@tarleton.edu>
}
